Selective policy enforcement (#57)
* Selective Policy Enforcement Enhancement

Initial draft for review

Signed-off-by: Ian Miller <imiller@redhat.com>

* Update enable mechanism to close race conditions

Signed-off-by: Ian Miller <imiller@redhat.com>

* Remove unneeded paragraphs

Signed-off-by: Ian Miller <imiller@redhat.com>

* Update cluster selection mechanism

Based on feedback and discussion update the mechanism for selecting
clusters for enforcement to use existing Placement APIs with optional
remediationActionOverride directives.

Signed-off-by: Ian Miller <imiller@redhat.com>

* Improve clarity of the new API description.

Signed-off-by: Ian Miller <imiller@redhat.com>

* Improve clarity of the new API description.

Signed-off-by: Ian Miller <imiller@redhat.com>

* Update based on comments.

Clarify behavior when there are multiple PlacementBindings.

Signed-off-by: Ian Miller <imiller@redhat.com>

* Move subFilter to top level.

The subFilter is a generic concept, not necessarily limited to the
remediationActionOverride. Moving it to the top level allows future
extensions of this functionality to other overrides/transforms which can
also make use of the subFilter.

Signed-off-by: Ian Miller <imiller@redhat.com>

* Add diagram for multiple PlacementBinding selection

Adding diagram depicting the selection logic for multiple
PlacementBindings with various subFilter and remediationActionOverride
settings. Courtesy of @JustinKuli

Signed-off-by: Ian Miller <imiller@redhat.com>

Signed-off-by: Ian Miller <imiller@redhat.com>