Revise OperatorPolicy design
The configuration of what upgrades are allowed has been changed from the
`endingCSV` field that was in `spec.subscription`, to a new field
`spec.allowedCSVs`. This makes the subscription field almost exactly
match the OLM subscription spec (only adding a `namespace` field), and
makes the behavior of the policy in `inform` mode more obvious.

The `spec.latestVersionNecessary` field has been brought into a new
section `spec.violationConditions` which will hold more configurations
of that type: things that *could* be cause for a violation, but not
necessarily for all users/policies.

The specification of the `status` is more complete now.

The compliance event now has an example YAML.

There are more User Stories, and they are generally smaller / more
specific, ideally each should be implementable in a reasonable amount
of time (some requiring other stories to already be completed).

Some details on approving InstallPlans has been added to the notes, with
observations from existing InstallPlans on clusters.

Refs:
 - https://issues.redhat.com/browse/ACM-3097

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>